Filter/requirements.txt
Flask
paho-mqtt
nats-py
python-dotenv
